city_app/pyproject.toml
[tool.poetry]
name = "city_app"
version = "0.1.0"
description = ""
authors = ["Egor Fironov <131647674+Egor1511@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "3.15.1"
psycopg2-binary = "^2.9.9"
geopy = "^2.4.1"
drf-yasg = "^1.21.7"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
setuptools = "^74.0.0"
django = "5.0.4"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
